outsource isempty of HPolyhedron
